requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pdfplumber==0.10.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
requests==2.31.0